pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# -----------------------
# Project metadata
# -----------------------
[project]
name = "accelerator-utilities"
version = "0.1.0"
description = "Framework-agnostic utilities and helpers for the accelerator framework"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "torch>=2.5.1",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "rich",
    "onnx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "mypy",
    "pre-commit",
]

# -----------------------
# Packaging (PEP 561)
# -----------------------
[tool.setuptools.packages.find]
where = ["src"]
include = ["accelerator*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
accelerator = ["py.typed"]

# -----------------------
# Black
# -----------------------
[tool.black]
line-length = 120
target-version = ["py39", "py310"]

# -----------------------
# Ruff (PEP 8 + PEP 257)
# -----------------------
[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

select = [
    # Core
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # bugbear

    # PEP 257
    "D",   # pydocstyle
]

ignore = [
    "E203",  # Black compatibility
    "E501",  # Handled by Black

    # pyupgrade
    "UP007", # X | Y not required yet
    "UP045", # X | Y not required yet

    # bugbear
    "B026",

    # Docstrings (pragmatic)
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public method
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public class
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D106",  # Missing docstring in public function
    "D107",  # Missing docstring in `__init__`
    "D205",  # 1 blank line required between summary line and description
    "D401",  # First line of docstring should be in imperative mood
]

[tool.ruff.isort]
combine-as-imports = true
known-first-party = ["accelerator"]

# -----------------------
# mypy (PEP 561)
# -----------------------
[tool.mypy]
python_version = "3.9"
strict = true
pretty = true
show_error_codes = true

namespace_packages = true
explicit_package_bases = true

mypy_path = "src"

exclude = [
    "tests/",
    "examples/",
]

# Relaxations commonly needed for infra / ML code
disallow_any_generics = false
warn_return_any = true
